create credentials file utils
