+++ tests/cli/test_database_sub_command.py
@@ -11,7 +11,7 @@ from uuid import uuid4
 from parameterized import parameterized
 
 from aito.api import create_table, delete_table, get_existing_tables, check_table_exists, query_entries, upload_entries
-from aito.cli.parser import get_credentials_file_config
+from aito.utils._credentials_file_utils import get_credentials_file_config
 from aito.schema import AitoTableSchema, AitoDatabaseSchema
 from tests.cli.parser_and_cli_test_case import ParserAndCLITestCase
 from tests.sdk.contexts import default_client, endpoint_methods_test_context
