+++ tests/cli/test_parser_utils.py
@@ -1,7 +1,6 @@
 from unittest.mock import patch
 
-from aito.cli.parser import ParseError, ArgParser, parse_env_variable, create_client_from_parsed_args, \
-    get_credentials_file_config, write_credentials_file_profile
+from aito.cli.parser import ParseError, ArgParser, parse_env_variable, create_client_from_parsed_args
 from aito.client import AitoClient, Error
 from tests.cases import BaseTestCase, CompareTestCase
 
@@ -11,52 +10,6 @@ class TestParserUtils(BaseTestCase):
         self.assertIsNone(parse_env_variable('RADIO_GA_GA'))
         with self.assertRaises(ParseError):
             parse_env_variable('RADIO_GA_GA', True)
-
-
-class TestCredentialsConfig(CompareTestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.input_folder = cls.input_folder.parent
-
-    def assertConfigProfile(self, config, profile_name, instance_url, api_key):
-        self.assertTrue(config.has_section(profile_name))
-        self.assertEqual(config.get(profile_name, 'instance_url'), instance_url)
-        self.assertEqual(config.get(profile_name, 'api_key'), api_key)
-
-    def test_get_config(self):
-        config = get_credentials_file_config(self.input_folder / 'sample_config')
-        self.assertEqual(config.sections(), ['default', 'space_oddity'])
-        self.assertConfigProfile(config, 'default', 'space_oddity', 'star_man')
-        self.assertConfigProfile(config, 'space_oddity', 'ground_control', 'major_tom')
-
-    def test_write_config(self):
-        out_fp = self.output_folder / 'nested' / 'config'
-
-        def clean_up_folder():
-            if out_fp.exists():
-                out_fp.unlink()
-            if out_fp.parent.exists():
-                out_fp.parent.rmdir()
-
-        self.addCleanup(clean_up_folder)
-        # write to a file with parents not created
-        write_credentials_file_profile('space_oddity', 'ground_control', 'major_tom', out_fp)
-        config = get_credentials_file_config(out_fp)
-        self.assertEqual(config.sections(), ['space_oddity'])
-        self.assertConfigProfile(config, 'space_oddity', 'ground_control', 'major_tom')
-        # write to a new profile
-        write_credentials_file_profile('star_man', 'dont_blow_it', 'lets_boogie', out_fp)
-        config = get_credentials_file_config(out_fp)
-        self.assertEqual(config.sections(), ['space_oddity', 'star_man'])
-        self.assertConfigProfile(config, 'space_oddity', 'ground_control', 'major_tom')
-        self.assertConfigProfile(config, 'star_man', 'dont_blow_it', 'lets_boogie')
-        # overwrite a profile
-        write_credentials_file_profile('star_man', 'heroes', 'just_for_one_day', out_fp)
-        config = get_credentials_file_config(out_fp)
-        self.assertEqual(config.sections(), ['space_oddity', 'star_man'])
-        self.assertConfigProfile(config, 'space_oddity', 'ground_control', 'major_tom')
-        self.assertConfigProfile(config, 'star_man', 'heroes', 'just_for_one_day')
 
 
 class TestCreateClientFromParsedArgs(CompareTestCase):
@@ -103,7 +56,7 @@ class TestCreateClientFromParsedArgs(CompareTestCase):
         self.stub_environment_variable('AITO_INSTANCE_URL', None)
         self.stub_environment_variable('AITO_API_KEY', None)
 
-        with patch('aito.cli.parser.DEFAULT_CREDENTIAL_FILE', self.input_folder / 'sample_config'):
+        with patch('aito.utils._credentials_file_utils.DEFAULT_CREDENTIAL_FILE', self.input_folder / 'sample_config'):
             self.assertEqual(
                 vars(AitoClient('space_oddity', 'star_man', False)),
                 vars(create_client_from_parsed_args(vars(self.parser.parse_args([])), check_credentials=False))
@@ -117,7 +70,7 @@ class TestCreateClientFromParsedArgs(CompareTestCase):
         )
         self.stub_environment_variable('AITO_INSTANCE_URL', None)
         self.stub_environment_variable('AITO_API_KEY', None)
-        with patch('aito.cli.parser.DEFAULT_CREDENTIAL_FILE', self.input_folder / 'sample_config'):
+        with patch('aito.utils._credentials_file_utils.DEFAULT_CREDENTIAL_FILE', self.input_folder / 'sample_config'):
             self.assertEqual(
                 vars(AitoClient('ground_control', 'major_tom', False)),
                 vars(create_client_from_parsed_args(expected_parsed_args, check_credentials=False))
@@ -127,7 +80,7 @@ class TestCreateClientFromParsedArgs(CompareTestCase):
         self.stub_environment_variable('AITO_INSTANCE_URL', None)
         self.stub_environment_variable('AITO_API_KEY', None)
         with self.assertRaises(ParseError):
-            with patch('aito.cli.parser.DEFAULT_CREDENTIAL_FILE', self.input_folder / 'sample_config'):
+            with patch('aito.utils._credentials_file_utils.DEFAULT_CREDENTIAL_FILE', self.input_folder / 'sample_config'):
                 create_client_from_parsed_args(vars(self.parser.parse_args(['--profile', 'random'])))
 
     def test_create_error_client(self):
